ADTnorm normalization to remove the technical variations across samples for each ADT marker.

This function removes the technical variations such as batch effect, sequencing depth biases, antibody selection difference and antibody concentration differences, etc. The normalized samples are ready for integration across studies.

ADTnorm(
  cell_x_adt = NULL,
  cell_x_feature = NULL,
  save_outpath = NULL,
  study_name = "ADTnorm",
  marker_to_process = NULL,
  bimodal_marker = NULL,
  trimodal_marker = NULL,
  positive_peak = NULL,
  bw_smallest_cd3 = 0.8,
  bw_smallest_cd4 = 0.8,
  bw_smallest_cd8 = 0.8,
  bw_smallest_bi = 1.1,
  bw_smallest_tri = 0.8,
  cd3_index = NULL,
  cd4_index = NULL,
  cd8_index = NULL,
  peak_type = "midpoint",
  multi_sample_per_batch = FALSE,
  shoulder_valley = FALSE,
  shoulder_valley_slope = -0.5,
  valley_density_adjust = 3,
  landmark_align_type = "negPeak_valley_posPeak",
  midpoint_type = "valley",
  neg_candidate_thres = asinh(8/5 + 1),
  lower_peak_thres = 0.001,
  brewer_palettes = "Set1",
  save_intermediate_rds = FALSE,
  save_intermediate_fig = TRUE,
  detect_outlier_valley = FALSE,
  target_landmark_location = NULL,
  clean_adt_name = FALSE
)

Arguments

cell_x_adt

Matrix of ADT raw counts in cells (rows) by ADT markers (columns) format.

cell_x_feature

Matrix of cells (rows) by cell features (columns) such as cell type, sample, and batch-related information.

save_outpath

The path to save the results.

study_name

Name of this run.

marker_to_process

Markers to normalize. Leaving empty to process all the ADT markers in cell_x_adt matrix.

bimodal_marker

Specify ADT markers that are likely to have two peaks based on researchers' prior knowledge or preliminary observation of particular data to be processed. Leaving it as default, ADTnorm will try to find the bimodal peak in all markers that are not listed in trimodal_marker.

trimodal_marker

Index of the ADT markers that tend to have three peaks based on researchers' prior knowledge (e.g., CD4) or preliminary observation on particular data to be processed.

positive_peak

A list variable containing a vector of ADT marker(s) and a corresponding vector of sample name(s) in matching order to specify that the uni-peak detected should be aligned to positive peaks. For example, for samples that only contain T cells. The only CD3 peak should be aligned to the positive peaks of other samples.

bw_smallest_cd3

The smallest bandwidth parameter value for CD3. Recommend 0.5 or 0.8.

bw_smallest_cd4

The smallest bandwidth parameter value for CD4. Recommend 0.5 or 0.8.

bw_smallest_cd8

The smallest bandwidth parameter value for CD8. Recommend 0.5 or 0.8.

bw_smallest_bi

The smallest bandwidth parameter value for bi-modal peaks. Recommend 1.1.

bw_smallest_tri

The smallest bandwidth parameter value for tri-modal peaks. Recommend the same value for CD4, such as 0.5.

cd3_index

Index for CD3 marker. If the ADT marker is named "CD3", the program can detect the index automatically. If users named CD3 uniquely, please provide the index for CD3 markers in the column name of cell_x_adt.

cd4_index

Index for CD4 marker. If the ADT marker is named "CD4", the program can detect the index automatically. If users named CD4 uniquely, please provide the index for CD4 markers in the column name of cell_x_adt.

cd8_index

Index for CD8 marker. If the ADT marker is named "CD8", the program can detect the index automatically. If users named CD8 uniquely, please provide the index for CD8 markers in the column name of cell_x_adt.

peak_type

The type of peak to be detected. Select from "midpoint" for setting the peak landmark to the midpoint of the peak region being detected or "mode" for setting the peak landmark to the mode location of the peak. "midpoint" can be generally more robust across samples and less impacted by the bandwidth. "mode" can be more accurate in determining the peak location if the bandwidth is generally ideal for the target marker.

multi_sample_per_batch

Set it to TRUE to discard the positive peak that only appear in one sample per batch (sample number is >=3 per batch).

shoulder_valley

Indicator to specify whether a shoulder valley is expected in case of the heavy right tail where the population of cells should be considered as a positive population.

shoulder_valley_slope

The slope on the ADT marker density distribution to call shoulder valley.

valley_density_adjust

Parameter for density function: bandwidth used is adjust*bw. This makes it easy to specify values like ‘half the default’ bandwidth.

landmark_align_type

Algin the peak and valleys using one of the "negPeak", "negPeak_valley", "negPeak_valley_posPeak", and "valley" alignment modes.

midpoint_type

Fill in the missing first valley by the midpoint of two positive peaks ("midpoint") or impute by other valleys ("valley").

neg_candidate_thres

The upper bound for the negative peak. Users can refer to their IgG samples to obtain the minimal upper bound of the IgG sample peak. It can be one of the values of asinh(4/5+1), asinh(6/5+1), or asinh(8/5+1) if the right 95% quantile of IgG samples is large.

lower_peak_thres

The minimal ADT marker density height of calling it a real peak. Set it to 0.01 to avoid a suspicious positive peak. Set it to 0.001 or smaller to include some small but tend to be real positive peaks, especially for markers like CD19.

brewer_palettes

Set the color scheme of color brewer.

save_intermediate_rds

Save the rds file for the intermediate objects.

save_intermediate_fig

Save the density plot figure for checking the peak and valley location detection.

detect_outlier_valley

Detect outlier valley and impute by the neighbor samples.

target_landmark_location

Align the landmarks to a fixed location or, by default, align to the mean across samples for each landmark. The default value is NULL. Setting it to "fixed" will align the negative peak to 1 and the right-most positive peak to 5. Users can also assign a two-element vector indicating the location of the negative and most positive peaks to be aligned.

clean_adt_name

Clean the ADT marker name

Examples

if (FALSE) {
ADTnorm(
  cell_x_adt = cell_x_adt,
  cell_x_feature = cell_x_feature,
  save_outpath = save_outpath,
  study_name = study_name,
  marker_to_process = c("CD3", "CD4", "CD8")
 )
}